+++ CHANGELOG.md
@@ -6,6 +6,13 @@ and this project tries to adhere to [Semantic Versioning](https://semver.org/spe
 
 ## [Unreleased]
 ### Changed
+- bus fleet distance is calculated dynamically from project start
+
+## [0.16.0] - 2020-07-15
+### Added
+- answered questions of finished category can be re-seen
+
+### Changed
 - merged landscape view into portrait view
 - app checks for mobile/desktop view and shows only related parts
 - major redesign of answer page (question form is shown with correct/wrong answers)